backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.30.0
alembic>=1.14.0
pydantic[email]>=2.6.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.18
redis>=5.2.0
httpx>=0.28.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.10
